refactor: 좌표와 관련된 클래스들은 coordinate 패키지로 이동
