+++ src/main/java/janggi/JanggiApplication.java
@@ -1,7 +1,7 @@
 package janggi;
 
 import janggi.board.Board;
-import janggi.board.Position;
+import janggi.coordinate.Position;
 import janggi.piece.Country;
 import janggi.view.InputView;
 import janggi.view.OutputView;
