+++ src/main/java/janggi/board/Board.java
@@ -1,5 +1,6 @@
 package janggi.board;
 
+import janggi.coordinate.Position;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
