+++ src/main/java/janggi/board/PieceInitialPosition.java
@@ -1,5 +1,6 @@
 package janggi.board;
 
+import janggi.coordinate.Position;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
