+++ src/main/java/janggi/piece/Cannon.java
@@ -1,7 +1,7 @@
 package janggi.piece;
 
 import janggi.board.Board;
-import janggi.board.Position;
+import janggi.coordinate.Position;
 import java.util.List;
 
 public class Cannon extends Piece {
