+++ src/main/java/janggi/piece/Chariot.java
@@ -1,7 +1,7 @@
 package janggi.piece;
 
 import janggi.board.Board;
-import janggi.board.Position;
+import janggi.coordinate.Position;
 import java.util.List;
 
 public class Chariot extends Piece{
