+++ src/main/java/janggi/view/InputView.java
@@ -1,6 +1,6 @@
 package janggi.view;
 
-import janggi.board.Position;
+import janggi.coordinate.Position;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
