+++ src/test/java/janggi/coordinate/PositionTest.java
@@ -1,8 +1,9 @@
-package janggi.board;
+package janggi.coordinate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import janggi.coordinate.Position;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
