+++ src/test/java/janggi/coordinate/RelativePositionTest.java
@@ -1,8 +1,9 @@
-package janggi.piece;
+package janggi.coordinate;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.board.Position;
+import janggi.coordinate.Position;
+import janggi.coordinate.RelativePosition;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
